CreditCardForm: Move non-components to helpers
This should make it easier to see where the component starts and ends.
